+++ gulpfile.js
@@ -5,9 +5,11 @@ const del = require('del')
 const Eleventy = require('@11ty/eleventy')
 const ssg = new Eleventy()
 const htmlmin = require('gulp-htmlmin')
+const validator = require('gulp-html')
 const gulpStylelint = require('gulp-stylelint')
 const sourcemaps = require('gulp-sourcemaps')
 const postcss = require('gulp-postcss')
+const sass = require('gulp-sass')
 const standard = require('gulp-standard')
 const webpack = require('webpack-stream')
 const compiler = require('webpack')
@@ -15,26 +17,24 @@ const path = require('path')
 const concat = require('gulp-concat')
 const beautify = require('gulp-beautify')
 const uglify = require('gulp-uglify')
+const svgmin = require('gulp-svgmin')
 const rename = require('gulp-rename')
 const connect = require('gulp-connect')
 
-const isProduction = config.get('node.environment') === 'production'
+const isProduction = config.get('eleventy.environment') === 'production'
 const SRC = config.get('paths.src.client')
 const BUILD = config.get('paths.build.client')
 
 const paths = {
-  liquid: {
-    src: `./${SRC}/**/*.liquid`
-  },
-  markdown: {
-    src: `./${SRC}/**/*.md`
-  },
   html: {
-    dest: `./${BUILD}`,
-    output: [
-      `./${BUILD}/**/*.html`,
-      `!./${BUILD}/courses/**/*.html`
-    ]
+    src: [
+      `./${SRC}/_data/**/*.*`, // Data
+      `./${SRC}/_includes/**/*.*`, // Includes
+      `./${SRC}/_layouts/**/*.*`, // Layouts
+      `./${SRC}/_content/**/*.*` // Content
+    ],
+    dest: `./${BUILD}`, // Build directory
+    output: `./${BUILD}/**/*.html` // Built HTML files
   },
   css: {
     all: `./${SRC}/_assets/css/**/*.css`,
@@ -43,9 +43,10 @@ const paths = {
     output: `./${BUILD}/css/bundle.css`
   },
   js: {
-    root: './*.js',
-    src: `./${SRC}/_assets/js/**/*.js`,
-    data: `./${SRC}/_data/**/*.js`,
+    src: [
+      './*.js',
+      `./${SRC}/_assets/js/**/*.js`
+    ],
     entry: {
       all: `./${SRC}/_assets/js/*.js`,
       index: `./${SRC}/_assets/js/index.js`
@@ -62,8 +63,12 @@ const paths = {
     dest: `./${BUILD}/img`
   },
   courses: {
-    src: `./${SRC}/courses/**/*`,
+    src: `./${SRC}/_content/courses/**/*`,
     dest: `./${BUILD}/courses`
+  },
+  svg: {
+    dest: `./${BUILD}/img`,
+    output: `./${BUILD}/img/**/*.svg`
   }
 }
 
@@ -77,169 +82,186 @@ async function html () {
   await ssg.init()
   await ssg.write()
 
-  const options = {
-    indent_size: 2,
-    max_preserve_newlines: 1
-  }
-
   const html = gulp.src(paths.html.output)
-    .pipe(beautify.html(options)) // Beautify
+    .pipe(beautify.html({
+      indent_size: 2,
+      max_preserve_newlines: 1
+    })) // Beautify
+    //
+    // TODO: inline critical CSS
+    // https://github.com/addyosmani/critical
+    // https://github.com/addyosmani/critical-path-css-demo/blob/dca7ec42c6b9d7bb2d8425c4055aabc753c1a6ac/gulpfile.js#L100-L111
+    //
     .pipe(gulp.dest(paths.html.dest))
     .pipe(connect.reload())
 
   return html
 }
 
+async function watchHtml () {
+  await ssg.init()
+  await ssg.watch()
+
+  const html = gulp.src(paths.html.output)
+    .pipe(beautify.html({
+      indent_size: 2,
+      max_preserve_newlines: 1
+    })) // Beautify
+    //
+    // TODO: inline critical CSS
+    // https://github.com/addyosmani/critical
+    // https://github.com/addyosmani/critical-path-css-demo/blob/dca7ec42c6b9d7bb2d8425c4055aabc753c1a6ac/gulpfile.js#L100-L111
+    //
+    .pipe(gulp.dest(paths.html.dest))
+    .pipe(connect.reload())
+
+  return html
+}
+
+function lint () {
+  const css = gulp.src(paths.css.src)
+    .pipe(gulpStylelint({
+      config: {
+        extends: ['stylelint-config-standard'],
+        rules: {
+          'at-rule-no-unknown': [true, {
+            ignoreAtRules: [
+              'include',
+              'mixin'
+            ]
+          }],
+          'no-descending-specificity': null,
+          'selector-pseudo-class-no-unknown': [true, {
+            ignorePseudoClasses: [
+              'focusring',
+              'readonly'
+            ]
+          }]
+        }
+      },
+      fix: true,
+      reporters: [
+        { formatter: 'string', console: true }
+      ]
+    }))
+
+  const js = gulp.src(paths.js.src)
+    .pipe(standard({ // Standard
+      fix: true,
+      envs: ['browser'] // https://eslint.org/docs/user-guide/configuring#specifying-environments
+    }))
+    .pipe(standard.reporter('default'))
+
+  const merged = merge(css, js)
+  return merged.isEmpty() ? null : merged
+}
+
 function css () {
-  const settings = {
-    config: {
-      extends: ['stylelint-config-standard'],
-      rules: {
-        'at-rule-no-unknown': [true, {
-          ignoreAtRules: ['include', 'mixin']
-        }],
-        'no-descending-specificity': null,
-        'selector-pseudo-class-no-unknown': null
-      }
-    },
-    fix: true,
-    reporters: [
-      { formatter: 'string', console: true }
-    ]
-  }
-
-  const plugins = [
-    require('postcss-easy-import'), // @import files
-    require('precss'), // Transpile Sass-like syntax
-    require('postcss-preset-env'), // Polyfill modern CSS
-    require('autoprefixer'), // Add vendor prefixes
-    require('pixrem')() // Add fallbacks for rem units
-  ]
-
-  const lint = gulp.src(paths.css.all)
-    .pipe(gulpStylelint(settings))
-
-  const build = gulp.src(paths.css.src)
+  return gulp.src(paths.css.src)
     .pipe(sourcemaps.init())
-    .pipe(postcss(plugins))
+    .pipe(sass().on('error', sass.logError)) // Preprocess Sass
+    .pipe(postcss([
+      require('postcss-easy-import'), // @import files
+      require('precss'), // Use Sass-like markup and staged CSS features
+      require('postcss-preset-env'), // Polyfill modern CSS
+      require('autoprefixer'), // Add vendor prefixes
+      require('pixrem')() // Add fallbacks for rem units
+    ]))
     .pipe(concat('bundle.css')) // Concatenate and rename
     .pipe(beautify.css({ indent_size: 2 })) // Beautify
     .pipe(sourcemaps.write('.')) // Maintain Sourcemaps
     .pipe(gulp.dest(paths.css.dest))
     .pipe(connect.reload())
-
-  const merged = merge(lint, build)
-
-  return merged.isEmpty() ? null : merged
 }
 
 function js () {
-  const options = {
-    // Standard configuration
-    fix: true,
-    envs: ['browser'] // https://eslint.org/docs/user-guide/configuring#specifying-environments
-  }
-
-  const config = {
-    // Webpack configuration
-    mode: isProduction ? 'production' : 'development',
-    entry: {
-      bundle: paths.js.entry.index
-    },
-    output: {
-      path: path.resolve(__dirname, 'build/js'),
-      publicPath: '/js/',
-      filename: '[name].js'
-    },
-    module: {
-      rules: [
-        {
-          test: /\.js$/,
-          exclude: /(node_modules)/,
-          use: {
-            loader: 'babel-loader', // Babel
-            options: {
-              "presets": [
-                [
-                  "@babel/preset-env",
-                  {
-                    "useBuiltIns": "entry"
-                  }
-                ]
-              ]
+  return gulp.src(paths.js.entry.all)
+    .pipe(sourcemaps.init())
+    .pipe(webpack({ // Bundle modules with Webpack
+      mode: isProduction ? 'production' : 'development',
+      entry: {
+        bundle: paths.js.entry.index
+      },
+      output: {
+        path: path.resolve(__dirname, 'build/js'),
+        publicPath: '/js/',
+        filename: '[name].js'
+      },
+      module: {
+        rules: [
+          {
+            test: /\.js$/,
+            exclude: /(node_modules)/,
+            use: {
+              loader: 'babel-loader', // Transpile JS with Babel
+              options: {
+                presets: ['@babel/preset-env']
+              }
             }
           }
-        }
-      ]
-    },
-    devtool: 'source-map'
-  }
-
-  const lint = gulp.src([paths.js.src, paths.js.root])
-    .pipe(standard(options))
-    .pipe(standard.reporter('default'))
-
-  const build = gulp.src(paths.js.entry.all)
-    .pipe(sourcemaps.init())
-    .pipe(webpack(config, compiler, err => {
+        ]
+      },
+      devtool: 'source-map'
+    }, compiler, err => {
       if (err) { throw new Error(err) }
-    })) // Webpack
+    }))
     .pipe(beautify({ indent_size: 2 })) // Beautify
     .pipe(sourcemaps.write('.'))
     .pipe(gulp.dest(paths.js.dest))
     .pipe(connect.reload())
-
-  const merged = merge(lint, build)
-
-  return merged.isEmpty() ? null : merged
 }
 
-function minify () {
-  const options = {
-    collapseBooleanAttributes: true,
-    collapseWhitespace: true,
-    decodeEntities: true,
-    html5: true,
-    minifyCSS: true,
-    minifyJS: true,
-    processConditionalComments: true,
-    removeAttributeQuotes: true,
-    removeComments: true,
-    removeEmptyAttributes: true,
-    removeOptionalTags: true,
-    removeRedundantAttributes: true,
-    removeScriptTypeAttributes: true,
-    removeStyleLinkTypeAttributes: true,
-    sortAttributes: true,
-    sortClassName: true,
-    trimCustomFragments: true,
-    useShortDoctype: true
-  }
-
-  const html = gulp.src(paths.html.output)
-    .pipe(htmlmin(options)) // Minify
+function minifyHtml () {
+  return gulp.src(paths.html.output)
+    .pipe(htmlmin({ // Minify
+      collapseBooleanAttributes: true,
+      collapseWhitespace: true,
+      decodeEntities: true,
+      html5: true,
+      minifyCSS: true,
+      minifyJS: true,
+      processConditionalComments: true,
+      removeAttributeQuotes: true,
+      removeComments: true,
+      removeEmptyAttributes: true,
+      removeOptionalTags: true,
+      removeRedundantAttributes: true,
+      removeScriptTypeAttributes: true,
+      removeStyleLinkTypeAttributes: true,
+      sortAttributes: true,
+      sortClassName: true,
+      trimCustomFragments: true,
+      useShortDoctype: true
+    }))
     .pipe(gulp.dest(paths.html.dest))
     .pipe(connect.reload())
+}
 
-  const css = gulp.src(paths.css.output)
+function minifyCss () {
+  return gulp.src(paths.css.output)
     .pipe(sourcemaps.init())
     .pipe(postcss([require('cssnano')])) // Minify
     .pipe(rename({ suffix: '.min' })) // Rename
     .pipe(sourcemaps.write('.')) // Maintain Sourcemaps
     .pipe(gulp.dest(paths.css.dest))
     .pipe(connect.reload())
+}
 
-  const js = gulp.src(paths.js.output)
+function minifyJs () {
+  return gulp.src(paths.js.output)
     .pipe(sourcemaps.init())
     .pipe(uglify()) // Minify
     .pipe(rename({ suffix: '.min' })) // Rename
     .pipe(sourcemaps.write('.')) // Maintain Sourcemaps
     .pipe(gulp.dest(paths.js.dest))
     .pipe(connect.reload())
+}
 
-  const merged = merge(html, css, js)
-
-  return merged.isEmpty() ? null : merged
+function minifySvg () {
+  return gulp.src(paths.svg.output)
+    .pipe(svgmin()) // Optimize and minify
+    .pipe(gulp.dest(paths.svg.dest))
+    .pipe(connect.reload())
 }
 
 function postMinify (cb) {
@@ -257,10 +279,12 @@ function postMinify (cb) {
 
 function assets () {
   const fonts = gulp.src(paths.fonts.src)
+    // TODO: Optimize fonts
     .pipe(gulp.dest(paths.fonts.dest))
     .pipe(connect.reload())
 
   const images = gulp.src(paths.images.src)
+    // TODO: Optimize images
     .pipe(gulp.dest(paths.images.dest))
     .pipe(connect.reload())
 
@@ -269,8 +293,37 @@ function assets () {
     .pipe(connect.reload())
 
   const merged = merge(fonts, images, courses)
-
   return merged.isEmpty() ? null : merged
+}
+
+function validate () {
+  return gulp.src(paths.html.output)
+    .pipe(validator({
+      'errors-only': true
+    })) // Validate HTML
+    .pipe(gulp.dest(paths.html.dest))
+}
+
+function watch (cb) {
+  gulp.watch(paths.html.src, gulp.series(watchHtml/*, validate*/)) // TODO: Uncomment validate
+  gulp.watch([paths.css.all], css)
+  gulp.watch(paths.js.src, js)
+  gulp.watch([
+    paths.fonts.src,
+    paths.images.src,
+    paths.courses.src
+  ], assets)
+
+  return cb()
+}
+
+function watchMinify (cb) {
+  gulp.watch([paths.html.output], minifyHtml)
+  gulp.watch([paths.css.output], gulp.series(minifyCss, postMinify))
+  gulp.watch([paths.js.output], gulp.series(minifyJs, postMinify))
+  gulp.watch([paths.svg.output], minifySvg)
+
+  return cb()
 }
 
 function serve (cb) {
@@ -279,42 +332,47 @@ function serve (cb) {
     livereload: true
   })
 
-  gulp.watch([
-    paths.js.data,
-    paths.liquid.src,
-    paths.markdown.src
-  ], html)
-  gulp.watch([paths.css.all], css)
-  gulp.watch([
-    paths.js.src,
-    paths.js.root
-  ], js)
-  gulp.watch([
-    paths.fonts.src,
-    paths.images.src,
-    paths.courses.src
-  ], assets)
-
-  cb()
+  return cb()
 }
 
 /**
  * Gulp tasks
  */
-const develop = gulp.series(
+exports.develop = gulp.series(
+  gulp.parallel(clean, lint),
+  gulp.parallel(html, css, js, assets),
+  validate
+)
+
+exports.serve = gulp.series(
+  gulp.parallel(clean, lint),
+  gulp.parallel(html, css, js, assets),
+  // validate, // TODO: Uncomment validate
+  watch,
+  serve
+)
+
+exports.watch = gulp.series(
+  gulp.parallel(clean, lint),
+  gulp.parallel(html, css, js, assets),
+  validate,
+  watch
+)
+
+exports.build = gulp.series(
   clean,
-  gulp.parallel(html, css, js),
-  assets
+  gulp.parallel(html, css, js, assets),
+  gulp.parallel(minifyCss, minifyJs, minifySvg),
+  gulp.parallel(minifyHtml, postMinify)
 )
 
-const build = gulp.series(
-  develop,
-  minify,
-  postMinify
+exports.production = gulp.series(
+  clean,
+  gulp.parallel(html, css, js, assets),
+  gulp.parallel(minifyCss, minifyJs, minifySvg),
+  gulp.parallel(minifyHtml, postMinify),
+  gulp.parallel(watch, watchMinify),
+  serve
 )
 
-exports.develop = develop
-exports.serve = gulp.series(develop, serve)
-exports.build = build
-exports.production = gulp.series(build, serve)
-exports.default = build
+exports.default = exports.build