+++ src/client/_content/pages/home.11tydata.js
@@ -1,0 +1,16 @@
+module.exports = {
+  title: 'Home',
+  slug: '.',
+  layout: 'home',
+  date: new Date('2020-02-20'),
+  api: '/api/agents',
+  resources: {
+    title: 'Additional Resources',
+    links: [
+      {
+        href: 'https://www.nar.realtor/fair-housing/fair-housing-month-at-home',
+        title: 'NAR Fair Housing at Home'
+      }
+    ]
+  }
+}