+++ src/client/_data/navigation.js
@@ -1,0 +1,10 @@
+const navigation = {
+  items: [
+    {
+      text: 'Home',
+      href: '/'
+    }
+  ]
+}
+
+module.exports = navigation