+++ src/server/handlers/api.js
@@ -1,0 +1,14 @@
+const config = require('config')
+
+module.exports = {
+  // Get home page index
+  getApi: (req, res, next) => {
+    res.render('index', {
+      title: config.get('api.title'),
+      routes: [
+        'pledges',
+        'users',
+      ]
+    })
+  }
+}