+++ src/server/middleware/auth.js
@@ -1,28 +1,50 @@
 const { log } = require('../modules/logger')
 const jwt = require('jsonwebtoken')
 const config = require('config')
+const debug = require('debug')('api:auth')
 
 module.exports = function (req, res, next) {
-  // Get a cookie
-  // response.cookies.x-auth-token
+  // Check headers
+  const origin = req.header('origin')
+  const referrer = req.header('referrer')
+  const url = config.get('api.url')
 
-  // or req.session
+  // If origin or referrer doesn't match, deny access
+  if (
+    (origin && !origin.includes(url)) ||
+    (referrer && !referrer.includes(url))
+  ) {
+    log.error('Access denied. Seems like a CSRF attack.', { status: 400 })
+    return res
+      .status(400)
+      .send('Access denied.')
+  }
 
-  // Most of CSRF attacks have a different origin or referrer header with your original host in their requests. So check if you have any of them in the header, are they coming from your domain or not! If not reject them.
+  // Check cookie
+  const token = req.cookies['x-auth-token']
 
-  const token = req.header('x-auth-token')
+  // If no token, deny access
   if (!token) {
     log.error('Access denied. No token provided.', { status: 401 })
     return res
       .status(401)
       .send('Access denied. No token provided.')
-      // TODO: Should we redirect to the login page?
-      // .redirect(401, '/login')
-      // Or let the client handle this?
   }
 
   try {
+    // Decode and verify the token signature
     const decoded = jwt.verify(token, config.get('jwtPrivateKey'))
+    const isExpired = Date.now() > decoded.exp
+
+    // If token is expired, deny access
+    if (isExpired) {
+      log.error('Access denied. Token is expired.', { status: 401 })
+      return res
+        .status(401)
+        .send('Access denied. Token is expired.')
+    }
+
+    // Set user to decoded token and continue
     req.user = decoded
     next()
   }