+++ CHANGELOG.md
@@ -15,6 +15,23 @@ and this project adheres to [Semantic Versioning](semver).
 ### Security
 -->
 
+## 1.1.0 - 2020-05-26 - Update build pipeline
+
+### Added
+- Add dependencies
+- Add GitHub workflow to ping Heroku
+
+### Changed
+- Update Readme
+- Update dependencies
+- Update server redirects and HTTP headers
+- Update build pipeline
+- Update environment variables
+- Update Eleventy configuration
+- Update GitHub Issue and Pull Request templates
+- Update server API - fresh NPSK clone, content type tweaks - TODO: Fix Pledges
+- Update client site files
+
 ## 1.0.0
 
 ### Added